+++ pkg/core/colors.go
@@ -1,0 +1,10 @@
+package core
+
+const (
+	Blue  = "\u001b[38;5;77m"
+	Green = "\u001b[38;5;45m"
+	Grey  = "\u001b[38;5;252m"
+	Red   = "\u001b[38;5;42m"
+	White = "\u001b[38;5;255m"
+	Reset = "\u001b[0m"
+)
